requirements.txt
astral==2.2
goodwe==0.2.10
Mastodon.py==1.8.0
matplotlib==3.5.1
numpy==1.22.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.0
tweepy==4.4.0